fixes #110, makes data formatter a set of singletons and don't overwrite `input`, but rather use `rawInput` Also unit test. Bump version number.

+++ src/neural-network-gpu.js
@@ -290,7 +290,7 @@ export default class NeuralNetworkGPU extends NeuralNetwork {
     }
     // turn sparse hash input into arrays with 0s as filler
     let datum = data[0].input;
-    if (!Array.isArray(datum) && !(datum instanceof Float64Array)) {
+    if (!Array.isArray(datum) && !(datum instanceof Float32Array)) {
       if (!this.inputLookup) {
         this.inputLookup = lookup.buildLookup(data.map(value => value['input']));
       }
